requirements.txt
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
beautifulsoup4==4.8.2
black==19.10b0
boto3
bs4==0.0.1
cachetools==4.0.0
certifi==2019.11.28
cfgv==3.1.0
chardet==3.0.4
click==7.1.1
coverage==5.0.4
coverage-badge==1.0.1
distlib==0.3.0
filelock==3.0.12
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
httplib2==0.17.1
identify==1.4.14
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
nodeenv==1.3.5
numpy==1.18.2
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.0.3
pathspec==0.7.0
pre-commit==2.2.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.4.4
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2019.3
PyYAML==5.3.1
-e git://github.com/quant-aq/py-quantaq.git#egg=quantaq
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
soupsieve==2.0
toml==0.10.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==20.0.17
wrapt==1.11.2
zipp==3.1.0
